se arreglo las ms inventario/ms reportes/ms restaurant
se modifico las cosas con la nueva bases de datos en todas las clases jpa

+++ ms-inventario/src/main/java/cl/hilton/inventario/dto/ProductoRequest.java
@@ -24,11 +24,11 @@ public class ProductoRequest {
 
     @NotNull
     @Min(0)
-    private Long stockActual;
+    private Integer stockActual;
 
     @NotNull
     @Min(0)
-    private Long stockMinimo;
+    private Integer stockMinimo;
 
     @NotBlank
     @Size(max = 20)
