+++ ms-reportes/src/main/java/cl/hilton/reportes/dto/ReporteResponse.java
@@ -1,4 +1,5 @@
 package cl.hilton.reportes.dto;
+
 import lombok.*;
 
 @Getter
@@ -8,7 +9,7 @@ import lombok.*;
 @Builder
 public class ReporteResponse {
 
-    private Integer id;
+    private Long id;
     private String codigo;
     private String nombre;
     private String descripcion;
